requirements.txt
PyQt5>=5.15
PyQtWebEngine>=5.15
google-generativeai>=0.6
